chore - split language status into two files
also fix https://github.com/microsoft/vscode/issues/231946 by replace icons with the spinner so that width stays constant
